Update Havemeyer 309 priority list.
Havemeyer 309 seating has the bottom part, and the balcony (top part).
The top part is more steep and easier to copy between rows, so we
really don't want seating to be dense there. Also it is more
disruptive to have TAs walk through it during the exam, and some side
parts of the balcony are totally impractical to access.

Assignment strategy:

- Fill every other seat in bottom part in columns (checkerboard
  pattern would allow cross-row copying)
- Every other seat of every other row in top part
- Double up students in the bottom part, starting in the front
- Previously skipped rows in the top part, center first
- Any remaining seats