Add Northwest Corner 501 seating chart.
I wrote two strategies for assigning seats.

The preferred strategy (noco501_ordered.txt) assigns seats from the
edges in. The idea is that it's easier for TAs to monitor students
near the staircases and the podium, so we should make the front/edges
dense before making the middle dense.  Concretely, this means:

- Assign every other seat in the front rows.
- Assign every other seat in the sides of the classroom.
- Assign all seats in the front.
- Assign all seats in the sides.
- Assign every other seat in the remaining areas of the classroom
  (middle and back).
- Assign all remaining seats.

The naive strategy (noco501_ordered_alt.txt) simply assigns all odd
seats, then all even seats. It is not recommended because it often
produces suboptimal seating charts for the reasons described above.

As usual, leftie desks are not included. According to @nsgomez --

"the desks on the right hand side (from the instructor desk looking
up) is leftie, with the exception of the A row. A12 is leftie but A17
is not."