seatingchart: Use argparse & working dir structure.
Each exam that you do will have a working directory under the out/
directory. The files under that working directory also have the working
directory's name appended to them. This ensures the filenames are not
ambiguous when emailing out seating charts (for example, if there are
two sections of a course).

By specifying the name of the working directory (the slug), the script
will automatically look in the out/ directory for files.

Reviewed by me