layouts: Add Schermerhorn 501.
Sch 501 has odd numbers on the left of the aisle, and evens on the
right. The leftie desks are the rightmost seat in each half of the
classroom, except the front row (A1 and A14).

There are also some movable desks and tables in the back of the
classroom (usually about 12 on each side of the room). I don't include
these in the seating chart since their positioning could change. Also,
they are not numbered.

I used the algorithm in 01d796c to decide the order. It's a little more
complicated because Sch 501 has an aisle going down the middle, which
makes the seats adjacent to the aisle easily accessible by TAs.